requirements.txt
streamlit
opencv-python-headless
numpy
tensorflow
sympy
pillow